Split HoF onto a separate page from the SDP. (#1389)
