+++ gatsby/src/components/HallOfFameEntry.js
